Renamed Abilities to Building folder; Moved BellCurve classes to its own folder; Added BellCurve Config file for default Life Resistance values;

+++ Assets/Scripts/Resistance/ResistanceUpgradeMessage.cs
@@ -4,5 +4,5 @@ using System.Collections;
 public class ResistanceUpgradeMessage
 {
     public R Type { get; internal set; }
-    public float Delta { get; internal set; }
+    public int Delta { get; internal set; }
 }
